requirements.txt
Django==5.2.5
djangorestframework==3.16.1
django-cors-headers==4.7.0
mysqlclient==2.2.7
Pillow==11.3.0
asgiref==3.9.1
sqlparse==0.5.3
tzdata==2025.2
